refactor: AdminPage에서 사용되는 tab들을 widgets로 이동
